commit from git

+++ DKPlugins/DK/DKObject.h
@@ -84,7 +84,7 @@ public:
 	static void Close(const DKString& id = ""){
 		DKDEBUGFUNC(id);		
 		//if(id.empty()){ return; }
-		for(unsigned int i = instances.size() - 1; i >= 0 && i < instances.size(); --i) {
+		for(unsigned long i = instances.size() - 1; i >= 0 && i < instances.size(); --i) {
 			if(id.empty() || same(id, instances[i]->data[1])){
 				//if(has(id,"/"))
 				//	DKERROR("Close(): "+id+" contains a /\n");
@@ -143,7 +143,7 @@ public:
 		DKDEBUGFUNC();
 		_instances = instances;
 	}
-	static int instance_count;
+	static unsigned long instance_count;
 private:
 	static std::vector<T*> instances;
 	static char* classname;
@@ -151,7 +151,7 @@ private:
 };
 
 template<class T, class R>	char* DKBaseT<T,R>::classname;
-template<class T, class R>	int DKBaseT<T,R>::instance_count;
+template<class T, class R>	unsigned long DKBaseT<T,R>::instance_count;
 template<class T, class R>	bool DKBaseT<T,R>::singleton;
 template<class T, class R>	std::vector<T*> DKBaseT<T,R>::instances;
 template <class T>
